requirements.txt
google-generativeai
fastapi
uvicorn[standard]
